I have reworked the Gulp file and how we are to compile and use assets in this project. I’ve set it up so that all assets are placed in the `src/main/assets` folder and then gulp will compile/copy the relevant assets over to the `src/main/public/assets` folder.
The `src/main/public/assets` folder should now be empty in the repository but running `npm run build-assets` (which is `gulp build`) this will:
- Copy the files into the public files folder
- Copy the fonts from the `ccs-frontend-kit` and `govuk-frontend` packages into the public fonts folder
- Copy the images from the `ccs-frontend-kit` and `govuk-frontend`packages and the images within the `src/main/assets/images` into the public images folder
- Copy the svgs from the `ccs-frontend-kit` package and the svg within the `src/main/assets/svg` into the public svg folder
- Compile the scripts from the `ccs-frontend-kit` and `govuk-frontend`packages and the scripts within the `src/main/assets/scripts` into a single file in the public scripts folder
- Compile the styles from within the `src/main/assets/styles` into a single file in the public styles folder

I’ve updated the travis file to build the assets before the deployment is done. This way we don’t need to have the compiled assets stored on the repo.

I’ve also created a new gulp task called `watch-assets` which can be run in parallel to the server while developing. This will automatically recompile the scripts/styles when changes are made within those folders. This means you don’t need to manually recompile the assets while developing.