As part of this commit I have been able to confirm that we do not use yarn at all (not in development, test or production), and so I have removed the scripts relating to yarn as they do not do anything.
In addition, I’ve sorted out the script for running the applications in development mode on Mac and Linux OS. To do this I make use of `ts-node-dev` which is a package designed exactly for this purpose.

As part of the previous refactor I have removed anything relating to webpack. Webpack is used for bundling assets and as we do that with gulp we do not need it. This allowed me to remove a lot of package imports and sort out some of the issues with the package imports.

Finally, I have reviewed the test scripts and make sure they can all run (even if the tests fail). The only outstanding issue is with the test coverage script which does not work to a memory size issue and will need to be fixed separately.