I have updated the configuration related to linting. Although linting related packages were added in the project, they were not used or not configured correctly.
I’ve now set it up to work with some basic rules. I’ve not told it to fix anything yet as there is something like 88,000 problems with the project.
There are separate scripts for linting the frontend and backend as, in its current state, it will be difficult to fix both at the same time.