I have removed express-nunjucks as we do not use it so don’t really need it as a dependency.
In addition, I’ve tidied up how we initialise Nunjucks so that it is more clear what is happening. Any additional changes are caused by VSCode formatting/linting the file when I save.

I’ve also updated the initialisation of Helmet too as just creating a new class instance and not assigning it is a strange thing to do as GC will just clear it up. Therefore, I’ve now made it functional.