Remove logging of the fetch response as it was clogging up the console.
This has come about due to the update due to an update to `@hmcts/nodejs-healthcheck` which updated `@hmcts/nodejs-logging` and this update changed the logging. Where as the previous version was happy to take and ignore non-string messages, this version does log the results (which are quite large) to the console.

As we don’t need to log it (it did not used to do anything), I’ve removed it from the code.