I have created services (code module) that can be used to handle API requests to the following services: - Public Procurement Gateway   - OAuth   - Organisation - Agreement Service - G-Cloud   - Service   - Supplier   - Search - Bank Holidays - Content Service
This means we won’t have to write the path we want to query in every instance of the code which should make it a lot more manageable.

I’ve also added tests for the endpoint to make sure the helpers I created do what we expect.

I also plan on doing this for the Tenders API but as it is quite large I want to do it separately. I have not done anything for the Logger instance as I am looking to remove that in the future. I have also not integrated any of the services with the app yet as I want to do that separately in the future.