It was pointed out recently that the choose agreements page takes a long time to load so I decided to investigate it. What I discovered is that we are making over 20 calls to the agreements service database and we are doing so synchronously.
Some of these requests are very slow and can cause a bottleneck when loading the page. After some testing I determined that it took the page, on average, about 7.5 seconds to do all the requests.

By making some changes to call the agreement API service more asynchronously I was able to get this down to an average of 3.4 seconds. This is obviously much better but is still pretty slow.

There are some requests to the agreement service that seem to take a little while (more than a second). We could also look at caching these requests as they seem to be made quite a lot.