Update the express session package and types and allow for any attribute in the session. This causes an issue with the deployments because of a slight change to how the `SessionData` type works in the latest version of `express-session`.
This caused an issue when deploying so I investigated further to find what the problem was.

I’m not sure why but the `npm run start` was not doing the type checking properly and missing definitions we had made in some declaration files. By adding the `--files` flag to the `.tsconfig` it makes sure that the declaration files are included and no type errors are thrown.

Another change I made was to update the `tsconfig.json` to have the full options and comments listed. I find this really useful to work out what each option does. I did not change any of the configurations.