Before I attempt to introduce some caching, I wanted to update redis to the latest version.
Because the connection to redis is no longer done synchronously, we need to await the redis connection before doing all the app setup. Therefore, I refactored the setup for the app to make it more clear what each part is doing.

After I did this I was able to update Redis to v4 and connect-redis to v7 (both the latest). Once the connection promise is resolved and Redis has been set up to be used for the session, we can then initialise app routes and middlewares.