After my endeavours on the choose agreements page, I’ve now taken a look at the dashboard page.
To start with, I created the services for the calls to the Tenders API that I need to make on this page. Because there are so many Tenders API endpoints my plan is to add them as we go along rather than try and work them all out at once.

I then inspected the dashboard middleware and did two major things initially. First of all, I refactored the calls to the Tenders API to make them asynchronous. Secondly I refactored the code that sorts and formats the response from the API to make it more clear what is doing (though I’m not sure I’ve had much effect on that).

After looking at the dashboard I then took a look at the search and found that there is a lot of repeated code and even more blocking API calls. I’ve also refactored that and moved it into the middleware. The way the page now works is there is one method to tender the dashboard and two middlewares, one for when there is no search and one for when there is a search.

Finally, I renamed a lot of things to make it more clear what exactly they are and also so they more align with TypeScript conventions.