While looking at some other work I decided to refactor the `genericFetch` method to include the logging of the performance automatically. It’s quite useful to log how long it takes to make an API request so we can workout performance bottlenecks. It’s also a good sanity check when writing asynchronous code that it does, in fact, execute asynchronously.
I’ve moved the logging and caching into their own files just to separate the concerns a little. I’ve removed the places in the code where we were logging the performance as they are no longer required. Finally, I have added tests to make sure that we do og as expected.