I’ve migrated our testing framework to [Vitest](https://vitest.dev/).
The test framework we currently use is Mocha but that requires configuration and making sure we add the right packages to get certain test functions (like mocking).

The other major player in the JS testing space is Jest but, from experience, this is not the best framework to use with TypeScript and ESModules.

Therefore, I’ve migrated the tests to Vitest as this has everything we need out of the box and uses TypeScript by default. Just to be clear, this is not some obscure framework and is linked to the VueJS framework (i.e. made by the same people).